1.1.0
Module now captures clipboard (Ctrl + V) events. If issues occur, this feature can be toggled on or off in settings.

When an image is pasted directly from a browser, the default name (image.png) is replaced with a random one to prevent browser caching and overwriting of existing file (image.webp).

Added option for GMs to select a specific folder for saving files.

Added option for GMs to retain original filenames